requirements.txt
asgiref==3.8.1
Django==4.2
django-admin-sortable2==2.1.10
django-tinymce==4.0.0
pillow==10.3.0
python-dotenv==1.0.1
sqlparse==0.5.0
typing_extensions==4.11.0
requests~=2.31.0